# Simplify GenericRateLimiter algorithm