--- a/util/rate_limiter.h
+++ b/util/rate_limiter.h
@@ -101,8 +101,8 @@ class GenericRateLimiter : public RateLimiter {
   Random rnd_;
 
   struct Req;
-  Req* leader_;
   std::deque<Req*> queue_[Env::IO_TOTAL];
+  bool waiting_;
 
   bool auto_tuned_;
   int64_t num_drains_;