--- a/util/rate_limiter.h
+++ b/util/rate_limiter.h
@@ -102,7 +102,7 @@ class GenericRateLimiter : public RateLimiter {
 
   struct Req;
   std::deque<Req*> queue_[Env::IO_TOTAL];
-  bool waiting_;
+  bool wait_until_refill_pending_;
 
   bool auto_tuned_;
   int64_t num_drains_;