@hx235 2021-08-04
Update
Done:
- understood the what/why/why-we-can of the simplification proposed by this PR (left some thoughts in the comments)
- left suggestion about PR summary
- left clarification questions for non-algorithmic changes

TODO:
- take another look for any concurrency issue (e.g, acceptable data race and non-acceptable data race)
- see if additional testing is needed
- take a closer read into the comment about the current implementation to see whether some of them (1) can be moved to PR summary (2)need further clarification