@hx235 2021-08-06
Review Summary

Done:
- Verified correctness of the new algorithm conceptually with my best efforts
    - By correctness, I mean: for every enqueued un-granted request, there exists an refill_interval where the request will be granted in a way that is subjected to rate limiting constraint (i.e, constraint on refill_byte_per_period and priority)
- Hand-traced as many execution orders as I could based on the structure of the new algorithm and verified that these execution orders do not violate correctness
- Left comments on the in-code comments and responded to reply from last review

TODO:
@ajkr - Other than suggestions related to structuring L148 to 183 into PR summary or in-line wherever you see fit (i.e, those marked with "suggestion") and a minor clarification [here](https://github.com/facebook/rocksdb/pull/8602/files#r684065178), the only coding related comment is [this](https://github.com/facebook/rocksdb/pull/8602/files#r683942787). Let me know if any further clarification/feedback is needed.

All others look good to me! Especially, based on my experience with learning both impl, compared with the existing impl, your simplified algorithm captures more explicitly an essential property to the correctness of our rate limiter algorithm (i.e, refill when time is right) and removes correctness-irrelevant property that was maintained in the leader election process of the existing impl (i.e, the front-queued property). This has made the code significantly shorter and the multi-thread code flow easier to draw. Good work 👍